+++ example/lib/app/pages/showcase/showcase_page.dart
@@ -8,6 +8,7 @@ import 'widgets/device_environment_card.dart';
 import 'widgets/logger_showcase_card.dart';
 import 'widgets/nano_command_card.dart';
 import 'widgets/navigation_showcase_card.dart';
+import 'widgets/repository_showcase_card.dart';
 import 'widgets/state_simulator_card.dart';
 import 'widgets/toast_showcase_card.dart';
 import 'widgets/universal_adapter_showcase_card.dart';
@@ -82,6 +83,8 @@ class _ShowcasePageState
               const LoggerShowcaseCard(),
               const SizedBox(height: 20),
               const UniversalAdapterShowcaseCard(),
+              const SizedBox(height: 20),
+              const RepositoryShowcaseCard(),
               const SizedBox(height: 20),
               NanoCommandCard(controller: controller),
             ],